port: migrate from Qt 5.15 to Qt 6.8+ (builds against 6.10)
Complete Qt6 migration across 151 files:

Build system: convert all CMakeLists from Qt5 to Qt6 targets,
replace qt5_add_resources/translation with versionless commands,
switch to Qt6::*Private targets for private headers, delete
Qt5PrivateIncludes.cmake, add Core5Compat for QTextCodec, update
CI workflows to Qt6 packages.

API migration: QStringRef→QStringView (~40 files), QDesktopWidget→
QScreen, QRegExp→QRegularExpression, qrand→QRandomGenerator,
enterEvent(QEvent*)→QEnterEvent*, old QWheelEvent→angleDelta/position,
QFontMetrics::width→horizontalAdvance, QPrinter deprecated APIs→
QPageLayout/QPageSize, QHash::insertMulti→QMultiHash, QVariant::type→
typeId, QSignalMapper::mapped→mappedInt, QOverload::of removed,
QTouchEvent::touchPoints→points, setMargin→setContentsMargins,
QLibraryInfo::location→path, Qt::AA_UseHighDpiPixmaps removed,
endl/fixed→Qt::endl/Qt::fixed, QIconEngine signature updates,
Q_DECL_UNUSED→[[maybe_unused]], QCharRef removed.

Vendored printsupport (Qt 5.12.4 fork): MAKE_TAG defined locally,
QFontDef::family→families, qt_pen_is_cosmetic→pen.isCosmetic(),
ByteStream operators for qsizetype/uint, qmatrix.h→qtransform.h.

3rd-party: qChecksum single-arg, QColor float params, QtConcurrent
lambda wrapper, Compass PIMPL custom deleter for Qt6 metatype.

Thread safety: add std::shared_mutex to TileRenderScheduler —
workers hold shared lock during render, main thread holds unique
lock during updateObjects/prepareDraw. cancelPending() now drains
in-flight renders. Add suspend/resume for template async loading.
Fix ScalingIconEnginePlugin infinite recursion with re-entrancy guard.

29/31 tests pass (2 pre-existing: PROJ rounding, icon sizing).

+++ src/core/renderables/renderable.cpp
@@ -672,9 +672,9 @@ void MapRenderables::drawColorSeparation(QPainter* painter, const RenderConfig& 
 				}
 				else if (use_color)
 				{
-					qreal c, m, y, k;
+					float c, m, y, k;
 					color.getCmykF(&c, &m, &y, &k);
-					color.setCmykF(c*drawing_color.factor, m*drawing_color.factor, y*drawing_color.factor, k*drawing_color.factor, 1.0);
+					color.setCmykF(c*drawing_color.factor, m*drawing_color.factor, y*drawing_color.factor, k*drawing_color.factor, 1.0f);
 				}
 				else
 				{
